+++ NEWS.md
@@ -8,5 +8,5 @@
 
 # SimSurvey 0.1.2.009
 
-* Simplify make_grid to minimize conflicts between strat splits and divisions
+* Simplify make_grid to minimize conflicts between strat splits and divisions; also add a "bezier" method to make_grid
 * Allow a vector of age-specific parameters to be supplied to sim_parabola plus add some options for defining a more asymetric parabola